Mongo

Mongo

API Reference

API Reference

Base.getproperty — Method

Overload property access as shortcut for collections.

db.project == db["project"]

Mongoc.find — Method

Create Mongoc.BSON arguments from pairs & invoke Mongoc.find

Pilr.bson — Method

bson(pair...) => Mongoc.BSON
bson(AbstractVector{pair}) => Vector{Mongoc.BSON}

Construct a BSON object using keyword arguments or pairs to reduce quote clutter.

Examples

julia> bson(:metadata!pt => r"^mei.*1$", :project => +:in => ["proj1", "proj2"])
Mongoc.BSON with 2 entries:
  "metadata.pt" => Dict{Any, Any}("\$regex"=>"^mei.*1\$")
  "project"     => Dict{Any, Any}("\$in"=>Any["proj1", "proj2"])

julia> bson([
         +:match => :type=>"SUBMIT", 
         +:group => (:_id=>+:pt, :N=>+:sum=>1)
       ])
Mongoc.BSON with 2 entries:
  "0" => Dict{Any, Any}("\$match"=>Dict{Any, Any}("type"=>"SUBMIT"))
  "1" => Dict{Any, Any}("\$group"=>Dict{Any, Any}("_id"=>"\$pt", "N"=>Dict{Any,…
julia> bson(a=1, b=2)
Mongoc.BSON with 2 entries:
  "a" => 1
  "b" => 2

Pilr.database — Function

database(jenkins_user [, db_name, [, db_password]] [localport = localport])

Construct a connected database defaults.

Pilr.database — Method

database(jenkins_user, db_name, db_password [, localport = 29030]) => Database

Return a Database connection.

Options

ssh_opts::AbstractVector{String}: extra arguments for the ssh tunnel command. Example: ["-Fnone", ""-i/home/me/.ssh/alt_rsa"]

Examples

julia> db = database("mmendel", "mei-s4r-qa", ENV["MONGO_PASSWORD"]) [...]

julia> import Mongoc

julia> Mongoc.count_documents(db["project"]) 1056

Pilr.dataset_collection — Function

dataset_collection(db, project_code, dataset_code, [ data | rawData | deleted ])

Return a Mongoc.Collection associated with a given PiLR dataset.

Examples

julia> db = database(ENV["JENKINS_USER"], QA, ENV["MONGO_PASSWORD"]);

Pilr.mfind — Method

New name for MongoDataFrames.pilrfind

Pilr.notifications — Method

Notifications that were scheduled when the notification time arrived. TODO: take LOGOUT into account

Pilr.parse_nginxlog — Method

parse_nginxlog(lines) => DataFrame

Creates a dataframe from nginx log file lines in NCSA Common Log format.

Examples

julia> lines = [
       "54.245.168.29 - - [26/Apr/2022:17:12:11 -0500] \"GET /login/auth HTTP/1.1\" 200 5603 \"-\" \"Amazon-Route53-Health-Check-Service (ref b1f0ca2a-3996-40b1-8f2e-6cf7267efa54; report http://amzn.to/1vsZADi)\"", 
       "128.148.225.62 - - [26/Apr/2022:17:12:12 -0500] \"POST /project/bluetooth_demo/emaOtsConfig/updateSurveyRules?config=82003&cardstack=Morning_Survey HTTP/1.1\" 302 0 \"https://cloud.pilrhealth.com/project/bluetooth_demo/emaOtsConfig/editSurveyRules?config=82003&cardstack=Morning_Survey\" \"Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/100.0.4896.127 Safari/537.36\""
       ];

julia> DataFrames.select(parse_nginxlog(lines), :status, :bytes, DataFrames.Not(:user))
2×7 DataFrame
 Row │ status  bytes  time                       remote_addr     request       ⋯
     │ UInt16  Int64  ZonedDat…                  SubString…      SubString…    ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │    200   5603  2022-04-26T17:12:11-05:00  54.245.168.29   GET /login/au ⋯
   2 │    302      0  2022-04-26T17:12:12-05:00  128.148.225.62  POST /project
                                                               3 columns omitted

Pilr.parse_tomcatlog — Method

parse_tomcatlog(stream) => (servletdf, rawdf)

There are 3 kinds of in the upstart/tomcat.log:

Output from the servlet loggers that start with timestamps like "2022-04-29 06:43:44,746"

Output from the tomcat server that start with timestamps like "Apr 29, 2022 9:50:48 AM"

Lines without a recognizable timestamp

The 1st kind are returned as rows in servletdf with all the fields parsed into columns. [TODO: describe columns] The rawdf has a row for each input line containing the original text plus...[TODO]

Example

julia> lines = [
       "2022-09-02 06:44:08,475 [307,ANONYMOUS,login/auth] INFO  pilrhealth.ActivityFilters  - >>",
       "2022-09-02 06:44:08,477 [307,ANONYMOUS,login/auth] INFO  pilrhealth.ActivityFilters  - << action processed in 2 millis"
       ];

julia> parse_tomcatlog(lines)
(2×8 DataFrame
 Row │ reqid     time                           user       action      level   ⋯
     │ Tuple…    ZonedDat…                      SubStrin…  SubStrin…   SubStri ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │ (307, 1)  2022-09-02T06:44:08.475-05:00  ANONYMOUS  login/auth  INFO    ⋯
   2 │ (307, 1)  2022-09-02T06:44:08.477-05:00  ANONYMOUS  login/auth  INFO
                                                               4 columns omitted, 2×4 DataFrame
 Row │ id     ref    time                           line                       ⋯
     │ Int64  Int64  ZonedDat…                      String                     ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │     1      1  2022-09-02T06:44:08.475-05:00  2022-09-02 06:44:08,475 [3 ⋯
   2 │     2      2  2022-09-02T06:44:08.477-05:00  2022-09-02 06:44:08,477 [3
                                                                1 column omitted)

Pilr.pilrDataFrame — Method

pilrDataFrame(db, project_code, dataset_code, [ (field=>value)... ]; [ sort ] [ limit ])

Fetch data froma PiLR dataset and convert it to a DataFrame with optional, common transformations.

A convenience function for invoking

Mongoc.find(dataset_collection(...), ...) |> flatdict |> DataFrame

with common default projections and conversions.

By default, it will project out the fields listed in DEFAULT_REMOVE and convert all DateTime fields to ZonedDateTime.

Arguments

db::Union{Pilr.Database,Mongoc.Database} - typically the result of Pilr.Database

`filter::Pair

Keyword arguments

The following options are passed along to Mongoc.find with bson automatically applied.

projection

sort

All other keyword arguments are passed on to Mongoc.find.

TODO

Convert all DateTime columns to ZonedDateTime. Have option to list fields that are actually local time.

Examples

julia> df = pilrDataFrame(database(ENV["JENKINS_USER"], QA, ENV["MONGO_PASSWORD"]),
                          "base_pilr_ema", APP_LOG,
                          "data.tag" => "SURVEY_QUEUE";
                          :sort=>:_id=>1, :limit=>1)
1×8 DataFrame
 Row │ timestamp                  metadata!pt  data!tag      data!msg          ⋯
     │ ZonedDat…                  String       String        String            ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │ 2018-01-10T18:39:34-06:00  pb1          SURVEY_QUEUE  Surveys displayed ⋯
                                                               5 columns omitted

Pilr.pilrShorten! — Method

Add a ZoneDateTime timestamp column to dataset DataFrame that combines metadata!timestamp and localTimestamp. Shorten path names

Pilr.pilrZonedTime — Function

pilrZonedTime(row, field = :metadata!timestamp) => ZonedDateTime
pilrZonedTime(dataframe, field = :metadata!timestamp) => Vector{ZonedDateTime}

Convert a Mongo date field to ZonedDateTime using the metadata!timestamp and localTimestamp to determine the offset.

Pilr.pilrZonedTime — Method

 pilrZonedTime(timestamp_with_offset) => ZonedDateTime

Parse a PiLR timestamp string to a ZonedDateTime.

julia> pilrZonedTime("2021-09-29T11:04:41-04:00")
2021-09-29T11:04:41-04:00

Pilr.remotelines — Method

remotelines(host, path; gunzip=true)

Create an iterator over lines in a remote file.

If the gunzip is true and the path ends with ".gz", it will be decompressed.

Intended for use with parse_tomcatlog and parse_nginxlog;

Pilr.surveyqueue — Method

surveyqueue(db, projectcode, field=>bsonquery... [ ; mongo_find_options...])

Create a dataframe from matching applog tag=SURVEYQUEUE entries.

Shows timestamped values of what a participant sees in their survey queue whenever EMA recalculates it.

Example

julia> surveyqueue(db, projcode, "metadata.pt"=>"278", "metadata.timestamp"=>"$gt"=>t)

Pilr.tomany — Method

tomany(parent, children...)

Pipeline helper TODO

Examples

Pilr.toparent — Method

toparent(key...; skipmissing=false)

Join with a chain of parents specified by the keys. o

Examples

pilrfind(db.participant, toparent(:project))
pilrfind(db.trigger. topparent(:configuration => :instrumentConfig))

Pilr.unflatten — Method

Convert flattened mongo doc to original shape

Pilr.DEFAULT_REMOVE — Constant

Columns that are moved to the end of returned DataFrames

Pilr.Database — Type

Wrap a Mongoc.Database and a Mongoc.Client

Pilr.MongoDataFrames.find — Method

find(collection [ , field => match_expr ]... ; [ , option => value ])

Short-hand syntax for invoking Mongoc.find and converting result to a DataFrame.

Example

julia> using Pilr, Pilr.MongoDataFrames, DataFrames

julia> db = database(ENV["JENKINS_USER"], QA, ENV["MONGO_PASSWORD"]);

julia> find(db["project"], :code=>+:regex=>"^test"; :limit=>2) |>
       df->select(df, :code, :dateCreated)
2×2 DataFrame
 Row │ code    dateCreated             
     │ String  DateTime                
─────┼─────────────────────────────────
   1 │ test1   2014-12-20T06:14:10.810
   2 │ test2   2015-01-12T17:28:50.481

Pilr.MongoDataFrames.pilrfind — Method

pilrfind(collection, queryitem...; limit=0, option...)
pilrfind(database, project_code, dataset_code, queritem...; limit=0, option...)

Query mongodb with a powerful short-hand syntax and return a dataframe.

Each queryitem can be a Pair or vector of pairs using bson syntax.

The vectors are flattened so you don't need ... after tomany[@ref].

If any item looks like a pipeline stage (starts with '"'), Mongoc.aggregate is called; otherwise Mongoc.find.

Examples

julia> db = database("mmendel",QA);

julia> pilrfind(db.project, :code=>+:regex=>"^test"; :limit=>1, :skip=>1)
1×12 DataFrame
 Row │ _id                       active  code    dateCreated              isDe ⋯
     │ String                    Bool    String  DateTime                 Bool ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │ 54b40452e4b098676ab9f366    true  test2   2015-01-12T17:28:50.481       ⋯
                                                               8 columns omitted
julia> configs = pilrfind(db.project, 
         :code=>+:regex=>"^test",
         tomany("project", "instrument", "instrumentConfig"),
         ; limit=3);

julia> configs[!, [:name, :instrumentConfig!name]]
3×2 DataFrame
 Row │ name            instrumentConfig!name 
     │ String          String                
─────┼───────────────────────────────────────
   1 │ Test Run        Protocol 1 Config
   2 │ Test James 102  test config #asdf
   3 │ Test Import     Branching Example