Mongo

Mongo

API Reference

API Reference

Pilr.bson — Function

bson(pair...) => Mongoc.BSON
bson(AbstractVector{pair}) => Vector{Mongoc.BSON}

Construct a BSON object using keyword arguments or pairs to reduce quote clutter.

Examples

julia> bson("metadata.pt" => "xyz", :project=>"\$in" => ["proj1", "proj2"])
Mongoc.BSON with 2 entries:
  "metadata.pt" => "xyz"
  "project"     => Dict{Any, Any}("\$in"=>Any["proj1", "proj2"])

julia> bson([
         "\$match"=>:type=>"SUBMIT", 
         "\$group"=>(:_id=>"\$pt", :N=>"\$sum"=>1)
       ])
Mongoc.BSON with 2 entries:
  "0" => Dict{Any, Any}("\$match"=>Dict{Any, Any}("type"=>"SUBMIT"))
  "1" => Dict{Any, Any}("\$group"=>Dict{Any, Any}("_id"=>"\$pt", "N"=>Dict{Any,…

julia> bson(a=1, b=2)
Mongoc.BSON with 2 entries:
  "a" => 1
  "b" => 2

Pilr.database — Function

database(jenkins_user, db_name, db_password [, localport = 29030]) => Database

Return a Database connection.

Options

ssh_opts::AbstractVector{String}: extra arguments for the ssh tunnel command. Example: ["-Fnone", ""-i/home/me/.ssh/alt_rsa"]

Examples

julia> db = database("mmendel", "mei-s4r-qa", ENV["MONGO_PASSWORD"]) [...]

julia> import Mongoc

julia> Mongoc.count_documents(db["project"]) 1056

database(jenkins_user [, db_name, [, db_password]] [localport = localport])

Construct a connected database defaults.

Pilr.dataset_collection — Function

dataset_collection(db, project_code, dataset_code, [ data | rawData | deleted ])

Return a Mongoc.Collection associated with a given PiLR dataset.

Examples

julia> db = database(ENV["JENKINS_USER"], QA, ENV["MONGO_PASSWORD"]);

Pilr.Database — Type

Wrap a Mongoc.Database and a Mongoc.Client

Pilr.flatdict — Function

flatdict(cursor ; [separator = "!"])

Convert an iterable of nested Dict{String,Any} (such as a Cursor returned by Mongoc.find) into a dictonary of equal length columns.

The returned dictionary can be converted to a DataFrame.

Each field value that is a dictionary is replaced a field for every entry. The field names are the path.

Using "!" does not require quoting in Symbol names, so you can type :metadata!pt instead of "metadata.pt".

Option Arguments

separator : path separator for flattened column names.

replace : either a vector of Pair{Symbol,Any} or a function (key, value) -> (key, value).

order : a vector of columns that should appear first.

Pilr.PilrDataFrame — Type

PilrDataFrame(db, project_code, dataset_code, [ filter... ]; [ sort ] [ limit ])

TODO

Convert all DateTime columns to ZonedDateTime. Have option to list fields that are actually local time.

Examples

julia> df = PilrDataFrame(database(ENV["JENKINS_USER"], QA, ENV["MONGO_PASSWORD"]),
                          "base_pilr_ema", APP_LOG,
                          "data.tag" => "SURVEY_QUEUE";
                          sort=:_id=>1, limit=1)
1×8 DataFrame
 Row │ timestamp                  metadata!pt  data!tag      data!msg          ⋯
     │ ZonedDat…                  String       String        String            ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │ 2018-01-10T18:39:34-06:00  pb1          SURVEY_QUEUE  Surveys displayed ⋯
                                                               5 columns omitted